+++ src/api/ResponseProcessor.ts
@@ -4,7 +4,7 @@ import { FitbitApiException } from '../exceptions';
 function tryParseJson(json: string): Record<string, any> | string {
     try {
         return JSON.parse(json);
-    } catch (exception) {
+    } catch {
         return json;
     }
 }