+++ src/constants/activity-types.ts
@@ -1,9 +1,11 @@
-export const BIKE = 90001;
-export const RUNNING = 90009;
-export const FENCING = 15200;
-export const WALKING = 90013;
-export const SKATEBOARDING = 15580;
-export const YOGA = 52000;
-export const SWIMMING = 90024;
-export const WEIGHT_TRAINING = 2131;
-export const CIRKUIT_TRAINING = 2040;
+export enum ActivityType {
+    BIKE = 90001,
+    RUNNING = 90009,
+    FENCING = 15200,
+    WALKING = 90013,
+    SKATEBOARDING = 15580,
+    YOGA = 52000,
+    SWIMMING = 90024,
+    WEIGHT_TRAINING = 2131,
+    CIRKUIT_TRAINING = 2040,
+}