+++ src/constants/index.ts
@@ -3,9 +3,4 @@ import * as INTRADAY_RESOURCES from './intraday-resources';
 import * as SCOPES from './scopes';
 import * as SUBSCRIPTION_COLLECTIONS from './subscription-collections';
 
-export {
-    ACTIVITY_TYPES,
-    INTRADAY_RESOURCES,
-    SCOPES,
-    SUBSCRIPTION_COLLECTIONS,
-};
+export { ACTIVITY_TYPES, INTRADAY_RESOURCES, SCOPES, SUBSCRIPTION_COLLECTIONS };