+++ src/constants/intraday-resources.ts
@@ -1,13 +1,8 @@
-export const CALORIES = 'calories';
-export const STEPS = 'steps';
-export const DISTANCE = 'distance';
-export const FLOORS = 'floors';
-export const ELEVATION = 'elevation';
-export const HEART = 'heart';
-
-export type IntradayResource = typeof CALORIES |
-    typeof STEPS |
-    typeof DISTANCE |
-    typeof FLOORS |
-    typeof ELEVATION |
-    typeof HEART;
+export enum IntradayResource {
+    CALORIES = 'calories',
+    DISTANCE = 'distance',
+    ELEVATION = 'elevation',
+    FLOORS = 'floors',
+    HEART = 'heart',
+    STEPS = 'steps',
+}