+++ src/constants/subscription-collections.ts
@@ -1,4 +1,6 @@
-export const ACTIVITIES = 'activities';
-export const BODY = 'body';
-export const FOODS = 'foods';
-export const SLEEP = 'sleep';
+export enum SubscriptionCollection {
+    ACTIVITIES = 'activities',
+    BODY = 'body',
+    FOODS = 'foods',
+    SLEEP = 'sleep',
+}