+++ src/types/api/ActivityFilters.ts
@@ -1,8 +1,0 @@
-import { DateTime } from 'luxon';
-
-export interface ActivityFilters {
-    beforeDate?: (DateTime | string),
-    afterDate?: (DateTime | string),
-    limit?: number,
-    offset?: number,
-}