+++ src/types/api/ApiActivityFilters.ts
@@ -1,0 +1,8 @@
+import { DateTime } from 'luxon';
+
+export interface ApiActivityFilters {
+    afterDate?: DateTime | string;
+    beforeDate?: DateTime | string;
+    limit?: number;
+    offset?: number;
+}