+++ src/types/api/ApiSleep.ts
@@ -1,0 +1,16 @@
+export interface ApiSleep {
+    dateOfSleep: string;
+    duration: number;
+    efficiency: number;
+    endTime: string;
+    infoCode: number;
+    levels: Record<string, unknown>;
+    logId: number;
+    minutesAfterWakeup: number;
+    minutesAsleep: number;
+    minutesAwake: number;
+    minutesToFallAsleep: number;
+    startTime: string;
+    timeInBed: number;
+    type: string;
+}