+++ src/types/api/ApiToken.ts
@@ -1,0 +1,9 @@
+export interface ApiToken {
+    access_token: string;
+    expireDate: string;
+    expires_in: number;
+    refresh_token: string;
+    scope: string;
+    token_type: string;
+    user_id: string;
+}