+++ src/types/api/Token.ts
@@ -1,9 +1,0 @@
-export interface Token {
-    access_token: string,
-    expires_in: number,
-    refresh_token: string,
-    scope: string,
-    token_type: string,
-    user_id: string,
-    expireDate: string,
-}