+++ src/types/api/index.ts
@@ -1,11 +1,5 @@
-import { Activity as ActivitySource } from './Activity';
-import { Sleep as SleepSource } from './Sleep';
-import { Token as TokenSource } from './Token';
-import { ActivityFilters as ActivityFiltersSource } from './ActivityFilters';
-import { DateFilters as DateFiltersSource } from './DateFilters';
-
-export type Activity = ActivitySource;
-export type Sleep = SleepSource;
-export type Token = TokenSource;
-export type ActivityFilters = ActivityFiltersSource;
-export type DateFilters = DateFiltersSource;
+export { ApiActivity } from './ApiActivity';
+export { ApiActivityFilters } from './ApiActivityFilters';
+export { ApiDateFilters } from './ApiDateFilters';
+export { ApiSleep } from './ApiSleep';
+export { ApiToken } from './ApiToken';